Добавление возможности сжигать фотографии

+++ code/modules/paperwork/paper.dm
@@ -300,29 +300,37 @@
 		\[hr\] : Adds a horizontal rule.
 	</BODY></HTML>"}, "window=paper_help")
 
-/obj/item/weapon/paper/proc/burnpaper(obj/item/weapon/lighter/P, mob/user)
-	var/class = "<span class='warning'>"
+
+/obj/item/weapon/proc/burnpaper(obj/item/weapon/lighter/P, mob/user) //weapon, to use this in paper_bundle and photo
+	var/list/burnable = list(/obj/item/weapon/paper,
+													/obj/item/weapon/paper_bundle,
+													/obj/item/weapon/photo)
+
+	if(!is_type_in_list(src, burnable))
+		return
 
 	if(P.lit && !user.restrained())
+		var/class = "<span class='warning'>"
 		if(istype(P, /obj/item/weapon/lighter/zippo))
 			class = "<span class='rose'>"
 
 		user.visible_message("[class][user] holds \the [P] up to \the [src], it looks like \he's trying to burn it!", \
 		"[class]You hold \the [P] up to \the [src], burning it slowly.")
 
-		spawn(20)
-			if(get_dist(src, user) < 2 && user.get_active_hand() == P && P.lit)
-				user.visible_message("[class][user] burns right through \the [src], turning it to ash. It flutters through the air before settling on the floor in a heap.", \
-				"[class]You burn right through \the [src], turning it to ash. It flutters through the air before settling on the floor in a heap.")
+		if(do_after(user, 20, TRUE, P, TRUE))
+			if((get_dist(src, user) > 1) || !P.lit)
+				return
+			user.visible_message("[class][user] burns right through \the [src], turning it to ash. It flutters through the air before settling on the floor in a heap.", \
+			"[class]You burn right through \the [src], turning it to ash. It flutters through the air before settling on the floor in a heap.")
 
-				if(user.get_inactive_hand() == src)
-					user.drop_from_inventory(src)
+			if(user.get_inactive_hand() == src)
+				user.drop_from_inventory(src)
 
-				new /obj/effect/decal/cleanable/ash(src.loc)
-				qdel(src)
+			new /obj/effect/decal/cleanable/ash(src.loc)
+			qdel(src)
 
-			else
-				user << "\red You must hold \the [P] steady to burn \the [src]."
+		else
+			user << "\red You must hold \the [P] steady to burn \the [src]."
 
 
 /obj/item/weapon/paper/Topic(href, href_list)
